Begin drafting structures for TSTP CNFs.
